requirements.txt
# Production dependencies
fastmcp>=2.13.0
notion-client>=2.6.0
python-dotenv>=1.1.1
mistune>=3.1.4